frameserver: deliver every discontinuity to the engine, locate ring drops, roll back a failed start
Registration actions were dispatched only for byte-carrying units, so the
discontinuity the classifier sets for holes, short and unframed units —
which never have a retained raster — never reached the engine, and its
temporal state crossed known byte discontinuities. Actions are now
dispatched for every classified observation; a unit shed for pool
exhaustion and any ring-full drops are also delivered as discontinuities.
Ring-full drops are folded into the next processed row's
preceding_ring_drops column so they are locatable in the timeline rather
than only counted, and the ingress count of eligible units is exposed as the
accounting denominator. fs_start rolls back a failed capture start instead
of stranding the worker; fs_close destroys the sync objects. The extended
unit counter stays 64-bit through the publisher. The default pool stays at
16 units: the whole-tape run reached a high-water of 2, and the size changes
only on a measured stall.

Tests: the fixture's hole must produce a discontinuity call; ingress equals
exact units when nothing is ring-dropped; a small-ring build (-DRING_ITEMS=2,
make test-smallring) with a stalled consumer must exhaust the ring, and the
preceding_ring_drops column must sum to the drops folded. Falsified: with
dispatch gated on bytes the hole never reaches the engine.

+++ src/frameserver/frame_publisher.c
@@ -62,7 +62,7 @@ int fp_open(fp_publisher **out, unsigned pool_size, const fp_sink *sink){
 }
 
 int fp_publish(fp_publisher *p, const uint8_t *unit, size_t unit_len,
-               uint32_t counter_ext, int d1, int d2, uint8_t transport){
+               uint64_t counter_ext, int d1, int d2, uint8_t transport){
     if (!p || !unit || unit_len != FP_UNIT_BYTES){ if (p) p->st.rejected_bad_args++; return -1; }
     IOSurfaceRef s = NULL;
     unsigned in_use = 0;
@@ -75,7 +75,7 @@ int fp_publish(fp_publisher *p, const uint8_t *unit, size_t unit_len,
     IOSurfaceLock(s, 0, NULL);
     fp_assemble((uint8_t *)IOSurfaceGetBaseAddress(s), unit, d1, d2);
     IOSurfaceUnlock(s, 0, NULL);
-    fp_frame f = { s, (uint64_t)counter_ext * 1001u, 30000u, counter_ext,
+    fp_frame f = { s, counter_ext * 1001u, 30000u, counter_ext,
                    (int8_t)clamp_offset(FP_FIELD1_START, d1), (int8_t)clamp_offset(FP_FIELD2_START, d2), transport };
     p->st.published++;
     p->sink.on_frame(p->sink.ctx, &f);
